Renamed BetterConfigEditor.vue to ConfigEditor.vue
